utils/requirements.txt
huggingface_hub
click
python-dotenv